# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/games": {
            "get": {
                "description": "Returns all games with players and winner info.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "List games",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (e.g., 'winner:Alice', 'player:Bob')",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Game"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new game with players; can optionally generate objectives.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Create a new game",
                "parameters": [
                    {
                        "description": "New game payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateGameInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "game, revealed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/games/assign-objective": {
            "post": {
                "description": "Admin action to attach an objective to a specific game round.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Manually assign a public objective to a round",
                "parameters": [
                    {
                        "description": "Assignment payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AssignObjectiveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/games/{game_id}/advance-round": {
            "post": {
                "description": "Advances the round; reveals a public objective unless none remain (then ends the game).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Advance game round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "game_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/games/{game_id}/speaker": {
            "post": {
                "description": "Assigns a speaker (initial or current) for a specific round.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Assign speaker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "game_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "player_id, round_id, is_initial",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/games/{id}": {
            "get": {
                "description": "Returns detailed game state with objective-based scoring breakdown.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get game detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently deletes a game by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Delete a game",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status and deleted game ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/games/{id}/exists": {
            "get": {
                "description": "Returns {\"exists\": true|false}.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Check game exists",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/games/{id}/objectives": {
            "get": {
                "description": "Returns all public objectives for a game, including stage and round.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get game public objectives",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/games/{id}/speaker/randomise": {
            "post": {
                "description": "Randomly selects a speaker for the game.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Randomise speaker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "speaker_id, speaker_name",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AssignObjectiveRequest": {
            "type": "object",
            "properties": {
                "game_id": {
                    "type": "integer"
                },
                "objective_id": {
                    "type": "integer"
                },
                "round_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.CreateGameInput": {
            "type": "object",
            "properties": {
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PlayerInput"
                    }
                },
                "speaker_id": {
                    "type": "integer"
                },
                "use_objective_decks": {
                    "type": "boolean"
                },
                "use_random_speaker": {
                    "type": "boolean"
                },
                "winning_points": {
                    "type": "integer"
                }
            }
        },
        "models.Game": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_round": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "game_number": {
                    "type": "integer"
                },
                "game_objectives": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GameObjective"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "partial": {
                    "type": "boolean"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GamePlayer"
                    }
                },
                "rounds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Round"
                    }
                },
                "speaker": {
                    "$ref": "#/definitions/models.Player"
                },
                "speakerAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SpeakerAssignment"
                    }
                },
                "speaker_id": {
                    "type": "integer"
                },
                "startingSpeakerID": {
                    "type": "integer"
                },
                "use_objective_decks": {
                    "type": "boolean"
                },
                "winner": {
                    "$ref": "#/definitions/models.Player"
                },
                "winner_id": {
                    "type": "integer"
                },
                "winning_points": {
                    "type": "integer"
                }
            }
        },
        "models.GameObjective": {
            "type": "object",
            "properties": {
                "GameID": {
                    "type": "integer"
                },
                "ID": {
                    "type": "integer"
                },
                "IsCDL": {
                    "type": "boolean"
                },
                "Objective": {
                    "$ref": "#/definitions/models.Objective"
                },
                "ObjectiveID": {
                    "type": "integer"
                },
                "Round": {
                    "$ref": "#/definitions/models.Round"
                },
                "RoundID": {
                    "type": "integer"
                },
                "Stage": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "revealed": {
                    "type": "boolean"
                }
            }
        },
        "models.GamePlayer": {
            "type": "object",
            "properties": {
                "faction": {
                    "type": "string"
                },
                "gameID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "playerID": {
                    "type": "integer"
                },
                "won": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Objective": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "stage": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.Player": {
            "type": "object",
            "properties": {
                "gameID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.PlayerInput": {
            "type": "object",
            "properties": {
                "faction": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Round": {
            "type": "object",
            "properties": {
                "gameID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Score"
                    }
                }
            }
        },
        "models.Score": {
            "type": "object",
            "properties": {
                "agendaTitle": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "gameID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "objective": {
                    "$ref": "#/definitions/models.Objective"
                },
                "objectiveID": {
                    "type": "integer"
                },
                "originallySecret": {
                    "type": "boolean"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "playerID": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "relicTitle": {
                    "type": "string"
                },
                "round": {
                    "$ref": "#/definitions/models.Round"
                },
                "roundID": {
                    "type": "integer"
                },
                "type": {
                    "description": "e.g. objective, imperial, support",
                    "type": "string"
                }
            }
        },
...
        "models.SpeakerAssignment": {
            "type": "object",
            "properties": {
                "game": {
                    "$ref": "#/definitions/models.Game"
                },
                "gameID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "player": {
                    "$ref": "#/definitions/models.Player"
                },
                "playerID": {
                    "type": "integer"
                },
                "round": {
                    "$ref": "#/definitions/models.Round"
                },
                "roundID": {
                    "type": "integer"
                }
            }
        }
    }
}